Change options for version prefix argument
PROBLEM
The second argument of `Area58check.encode/2` are options, which
currently is only the version which is mandatory. I don't think this is
the right thing knowing that the version has to be stated. Plus there
are no current options at all.

SOLUTION
Now the second parameter is the version. Updated the documentation of
the type to put that clear.